requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==2.3.7
python-dotenv